+++ lua/personal/utils/prequire.lua
@@ -1,0 +1,7 @@
+---protected require call
+---@param m string
+---@return any module the module itself. nil if not found.
+return function(m)
+	local _, err = pcall(require, m)
+	return err
+end