+++ lua/plugins/config/focus_config.lua
@@ -4,6 +4,7 @@ return function(use)
 		config = function()
 			require('focus').setup({
 				signcolumn = false, -- disable sign column only on main window
+				compatible_filetrees = { 'nvimtree', 'nerdtree', 'chadtree', 'fern', 'Calltree' },
 			})
 		end,
 	})